requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
python-docx>=1.1.0
PyPDF2>=3.0.0
pandas>=2.1.0
plotly>=5.18.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
